Added an include for stdlib.
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6097 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ngmath/src/examples/enatgrid/c_nnex01d.c
@@ -24,7 +24,7 @@ main()
   float outd[NUMXOUT * NUMYOUT], xp[NUMXOUT], yp[NUMYOUT];
   Gcolr_rep rgb;
 
-  iwk = (int *) calloc(2*NUMXOUT*NUMYOUT,sizeof(float));
+  iwk = (int *) calloc(2*NUMXOUT*NUMYOUT,sizeof(int));
 
   xc = 1./(NUMXOUT-1.);
   for (i = 0 ; i < NUMXOUT ; i++) {
